bluetooth: remove generation of random passkey in default agent
The default agent implements default functions that are meant to be
overriden by application-specific implementation. The default
implementation of "request passkey" in the agent would generate a random
passkey, which is not what it is supposed to do. Instead the
application-specific implementation should override this function and get the
passkey input from the user, as implemented in the agent test program.
The default function should never be called in a product, therefore it
has been modified to return an error instead of the random passkey.

Change-Id: I01e766885fff56bec5f5de0575d69af12c58b192
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>