wifi: fix handling of non ASCII SSID names
In some cases, a Wifi access point can expose a SSID containing non
ASCII characters. In this case wpa-supplicant encodes the SSID name in
hexa in the returned SSID string. This patchs detects this situation and
properly decodes the SSID name before storing it into the acess point
information structure.

Change-Id: Ida74340adbd81aa19f4ba136a74ad8b80b130457
Signed-off-by: Myunghoon Ahn <myunghoon.ahn@samsung.com>
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>