serial: fix null pointer dereferencing and potential memory leak
The C++ layer of the serial module wrongfully handled the name string
in the configuration structure, and did not check for null pointer
passed by the calling application when duplicating the string. This
patch gets rid of these potential segfaults and memory leak due to the
destructor not freeing the duplicated string.

Change-Id: Ib9890a4d7634eede6deac1e47abfe4bab8592ff6
Signed-off-by: Gregory Lemercier <g.lemercier@samsung.com>