cloud: Fix NULL checking bug of ssl_config parameter
From v1.4.0, parameter of websocket_open_stream() is modified.
Especially artik_ssl_config *ssl_config is added, and this can be NULL in doxygen.
But in the code, there is no NULL checking point, so if user puts
this parameter as NULL, "Segmentation fault" is occurred.
Moreover code flows do not consider NULL parameter, so if the input parameter is NULL,
some initializations should be required.
This fixes this bug.

Change-Id: I3967e5b1d5cab52fe43b408e81aaad66583ac577
Signed-off-by: Joonghong Park <joonghong.park@samsung.com>